PDF to Excel Converter/backend/requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.0
colorama==0.4.6
cryptography==44.0.3
et_xmlfile==2.0.0
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pdf2image==1.17.0
pdfminer.six==20250327
pdfplumber==0.11.6
pillow==11.2.1
pycparser==2.22
pypdfium2==4.30.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tzdata==2025.2
Werkzeug==3.1.3
